extract propagation bridge helpers
